+++ src/spreadsheet/SpreadsheetViewportWidget.js
@@ -16,7 +16,7 @@ import SpreadsheetLabelName from "./reference/SpreadsheetLabelName.js";
 import SpreadsheetMessenger from "./message/SpreadsheetMessenger.js";
 import SpreadsheetMessengerCrud from "./message/SpreadsheetMessengerCrud.js";
 import SpreadsheetMetadata from "./meta/SpreadsheetMetadata.js";
-import SpreadsheetViewport from "./reference/SpreadsheetViewport.js";
+import SpreadsheetViewport from "./SpreadsheetViewport.js";
 import Table from '@material-ui/core/Table';
 import TableBody from '@material-ui/core/TableBody';
 import TableCell from '@material-ui/core/TableCell';
